refactor: organize folder
