+++ services/transactions/src/interfaces/public-api/transactions-public-api.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TransactionsModule } from '../../transactions.module.js';
-import { TransactionsController } from './transactions-controller.js';
-import { BalanceController } from './balance-controller.js';
+import { TransactionsController } from './transactions/transactions-controller.js';
+import { BalanceController } from './balance/balance-controller.js';
 
 @Module({
   imports: [TransactionsModule],
