+++ services/transactions/src/interfaces/public-api/transactions/list-transactions.request.dto.ts
@@ -1,5 +1,5 @@
 import { ApiPropertyOptional } from '@nestjs/swagger';
-import { TransactionType } from '../../domain/transaction-type.enum.js';
+import { TransactionType } from '../../../domain/transaction-type.enum.js';
 import { IsEnum, IsOptional } from 'class-validator';
 
 export class ListTransactionsRequestDto {
