+++ services/transactions/src/interfaces/public-api/transactions/transaction.response.dto.ts
@@ -1,9 +1,9 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { TransactionType } from '../../domain/transaction-type.enum.js';
+import { TransactionType } from '../../../domain/transaction-type.enum.js';
 import { Optional } from '@nestjs/common';
 import { IsEnum, IsInt, IsNotEmpty, IsNumber, Min } from 'class-validator';
 
-export class TransactionModelDto {
+export class TransactionResponseDto {
   @ApiProperty()
   @Optional()
   id?: string;
